#116 organize files a bit more obvious
