+++ client/src/Books.js
@@ -1,7 +1,7 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 'use strict';
 
-var Link = require("./Link.js");
+var Link = require("./router/Link.js");
 var CssJs = require("bs-css-emotion/src/CssJs.js");
 var Curry = require("rescript/lib/js/curry.js");
 var React = require("react");
@@ -10,7 +10,7 @@ var $$Request = require("rescript-request/src/Request.js");
 var Belt_Option = require("rescript/lib/js/belt_Option.js");
 var Caml_option = require("rescript/lib/js/caml_option.js");
 var Belt_SetString = require("rescript/lib/js/belt_SetString.js");
-var ResponseMapper = require("./ResponseMapper.js");
+var ResponseMapper = require("./http/ResponseMapper.js");
 var StringToColor = require("string-to-color");
 
 function str(s) {
