+++ client/src/CssJsBugBorderRadius4.js
@@ -1,0 +1,13 @@
+// Generated by ReScript, PLEASE EDIT WITH CARE
+'use strict';
+
+var CssJs = require("bs-css-emotion/src/CssJs.js");
+
+var btnOutlinePrimary = CssJs.style([CssJs.borderRadius4(CssJs.px(0), CssJs.px(0), CssJs.px(2), CssJs.px(1))]);
+
+var Styles = {
+  btnOutlinePrimary: btnOutlinePrimary
+};
+
+exports.Styles = Styles;
+/* btnOutlinePrimary Not a pure module */
