+++ client/src/Navbar.js
@@ -1,7 +1,7 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 'use strict';
 
-var Link = require("./Link.js");
+var Link = require("./router/Link.js");
 var CssJs = require("bs-css-emotion/src/CssJs.js");
 var React = require("react");
 var Css_Js_Core = require("bs-css/src/Css_Js_Core.js");
@@ -52,7 +52,7 @@ var Styles = {
   icon: icon
 };
 
-function Header(Props) {
+function Navbar(Props) {
   return React.createElement("div", {
               className: CssJs.merge([
                     fontStyles,
@@ -70,7 +70,7 @@ function Header(Props) {
                 }));
 }
 
-var make = Header;
+var make = Navbar;
 
 exports.Styles = Styles;
 exports.make = make;
