--- client/src/StringToColor.js
@@ -1,2 +1,0 @@
-// Generated by ReScript, PLEASE EDIT WITH CARE
-/* This output is empty. Its source's type definitions, externals and/or unused code got optimized away. */
