--- client/src/StringToColor.res
@@ -1,1 +1,0 @@
-@module external toHex: string => string = "string-to-color"
